+++ Sources/Core/GameUpdatable.swift
@@ -1,5 +1,5 @@
 import Foundation
 
 public protocol GameUpdatable {
-    func update(input: Input, delta: Double)
+    func update(input: Input, delta: Float)
 }
